# openapi.yaml
openapi: 3.0.3
info:
  title: Appraisal System API
  description: REST API for an Employee Appraisal Management System built with Flask and MongoDB.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development server
...
paths:
...
  /api/appraisals/:
    post:
      tags: [Appraisal]
      summary: Create a new appraisal
      description: >
        Creates an appraisal for the given employee. The reviewer is
        automatically set from the logged-in user's session.
      operationId: createAppraisal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppraisalRequest"
      responses:
        "201":
          description: Appraisal created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Appraisal created successfully
                  appraisal:
                    $ref: "#/components/schemas/Appraisal"
        "400":
          description: Missing required field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      tags: [Appraisal]
      summary: Get all appraisals
      description: Returns a list of appraisals with optional filters.
      operationId: getAllAppraisals
      parameters:
        - name: employee_id
          in: query
          description: Filter by employee ID
          required: false
          schema:
            type: string
        - name: status
          in: query
          description: Filter by appraisal status
          required: false
          schema:
            $ref: "#/components/schemas/AppraisalStatus"
      responses:
        "200":
          description: List of appraisals
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 3
                  appraisals:
                    type: array
                    items:
                      $ref: "#/components/schemas/Appraisal"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/appraisals/{appraisal_id}:
    get:
      tags: [Appraisal]
      summary: Get a specific appraisal
      operationId: getAppraisal
      parameters:
        - $ref: "#/components/parameters/AppraisalIdPath"
      responses:
        "200":
          description: Appraisal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appraisal"
        "400":
          description: Invalid appraisal ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Appraisal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      tags: [Appraisal]
      summary: Update an appraisal
      description: >
        Updates the specified appraisal. Only the allowed fields can be
        modified; employee_id and reviewer info cannot be changed.
      operationId: updateAppraisal
      parameters:
        - $ref: "#/components/parameters/AppraisalIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppraisalRequest"
      responses:
        "200":
          description: Appraisal updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Appraisal updated successfully
                  appraisal:
                    $ref: "#/components/schemas/Appraisal"
        "400":
          description: Invalid appraisal ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Appraisal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags: [Appraisal]
      summary: Delete an appraisal
      operationId: deleteAppraisal
      parameters:
        - $ref: "#/components/parameters/AppraisalIdPath"
      responses:
        "200":
          description: Appraisal deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Appraisal deleted successfully
        "400":
          description: Invalid appraisal ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Appraisal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/appraisals/{appraisal_id}/status:
    patch:
      tags: [Appraisal]
      summary: Update appraisal status
      description: >
        Transitions an appraisal through its workflow statuses
        (draft → submitted → approved / rejected).
      operationId: updateAppraisalStatus
      parameters:
        - $ref: "#/components/parameters/AppraisalIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: "#/components/schemas/AppraisalStatus"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Appraisal status updated successfully
                  appraisal:
                    $ref: "#/components/schemas/Appraisal"
        "400":
          description: Invalid appraisal ID or invalid status value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Appraisal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/appraisals/my/appraisals:
    get:
      tags: [Appraisal]
      summary: Get my appraisals
      description: Returns appraisals where the logged-in user is the employee being appraised.
      operationId: getMyAppraisals
      responses:
        "200":
          description: List of the current user's appraisals
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  employee_id:
                    type: string
                    example: EMP001
                  appraisals:
                    type: array
                    items:
                      $ref: "#/components/schemas/Appraisal"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/appraisals/my/reviews:
    get:
      tags: [Appraisal]
      summary: Get my reviews
      description: Returns appraisals where the logged-in user is the reviewer.
      operationId: getMyReviews
      responses:
        "200":
          description: List of appraisals reviewed by the current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  reviewer_id:
                    type: string
                    example: EMP001
                  appraisals:
                    type: array
                    items:
                      $ref: "#/components/schemas/Appraisal"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/appraisals/my/summary:
    get:
      tags: [Appraisal]
      summary: Get my appraisal summary
      description: Returns dashboard summary metrics for the logged-in user.
      operationId: getMySummary
      responses:
        "200":
          description: Summary metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyAppraisalSummary"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/appraisals/my/review-queue:
    get:
      tags: [Appraisal]
      summary: Get my review queue
      description: Returns appraisals assigned to the current reviewer for a given status.
      operationId: getMyReviewQueue
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AppraisalStatus"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        "200":
          description: Review queue list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyReviewQueueResponse"
        "400":
          description: Invalid filter or limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/appraisals/employee/{employee_id}/history:
    get:
      tags: [Appraisal]
      summary: Get employee appraisal history
      description: Returns latest appraisals for a specific employee.
      operationId: getEmployeeAppraisalHistory
      parameters:
        - $ref: "#/components/parameters/EmployeeIdPath"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 12
      responses:
        "200":
          description: Appraisal history list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppraisalHistoryResponse"
        "400":
          description: Invalid limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/appraisals/{appraisal_id}/submit:
    post:
      tags: [Appraisal]
      summary: Submit appraisal
      description: Submits a draft appraisal for review approval.
      operationId: submitAppraisal
      parameters:
        - $ref: "#/components/parameters/AppraisalIdPath"
      responses:
        "200":
          description: Appraisal submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppraisalWriteResponse"
        "400":
          description: Invalid appraisal ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Current user cannot submit this appraisal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Appraisal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Appraisal is not in draft status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/appraisals/{appraisal_id}/decision:
    patch:
      tags: [Appraisal]
      summary: Approve or reject appraisal
      description: Applies reviewer decision to a submitted appraisal.
      operationId: decideAppraisal
      parameters:
        - $ref: "#/components/parameters/AppraisalIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppraisalDecisionRequest"
      responses:
        "200":
          description: Appraisal decision applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppraisalWriteResponse"
        "400":
          description: Invalid appraisal ID or decision payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Current user cannot decide this appraisal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Appraisal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Appraisal is not in submitted status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"

# ════════════════════════════════════════════════════════════════════════
# Components
# ════════════════════════════════════════════════════════════════════════
components:
...
  parameters:
    EmployeeIdPath:
      name: employee_id
      in: path
      required: true
      description: Auto-generated employee ID (e.g. EMP001)
      schema:
        type: string
        example: EMP001
    AppraisalIdPath:
      name: appraisal_id
      in: path
      required: true
      description: MongoDB ObjectId of the appraisal
      schema:
        type: string
        example: 507f1f77bcf86cd799439011

  # ── Schemas ──────────────────────────────────────────────────────────
  schemas:
...
    AppraisalStatus:
      type: string
      enum:
        - draft
        - submitted
        - approved
        - rejected
      example: draft

    # ─ Employee ─
...
    Appraisal:
      type: object
      description: Serialized appraisal document
      properties:
        id:
          type: string
          description: MongoDB ObjectId
          example: 60d0fe4f5311236168a109cb
        employee_id:
          type: string
          example: EMP001
        employee_name:
          type: string
          example: John Doe
        reviewer_id:
          type: string
          example: EMP002
        reviewer_name:
          type: string
          example: Jane Smith
        appraisal_period:
          type: string
          description: "Period being appraised, e.g. Q1 2026"
          example: Q1 2026
        performance_rating:
          type: integer
          description: Rating on a 1-5 scale
          minimum: 0
          maximum: 5
          example: 4
        goals_achieved:
          type: array
          items:
            type: string
          example:
            - Completed project X
            - Improved test coverage
        strengths:
          type: string
          example: Strong technical skills
        areas_for_improvement:
          type: string
          example: Time management
        comments:
          type: string
          example: Overall good performance
        status:
          $ref: "#/components/schemas/AppraisalStatus"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ─ Request Bodies ─
...
    CreateAppraisalRequest:
      type: object
      required:
        - employee_id
        - appraisal_period
      properties:
        employee_id:
          type: string
          example: EMP001
        appraisal_period:
          type: string
          example: Q1 2026
        performance_rating:
          type: integer
          minimum: 0
          maximum: 5
          default: 0
        goals_achieved:
          type: array
          items:
            type: string
          default: []
        strengths:
          type: string
          default: ""
        areas_for_improvement:
          type: string
          default: ""
        comments:
          type: string
          default: ""
        status:
          $ref: "#/components/schemas/AppraisalStatus"
    UpdateAppraisalRequest:
      type: object
      description: Only the fields provided will be updated.
      properties:
        appraisal_period:
          type: string
        performance_rating:
          type: integer
          minimum: 0
          maximum: 5
        goals_achieved:
          type: array
          items:
            type: string
        strengths:
          type: string
        areas_for_improvement:
          type: string
        comments:
          type: string
        status:
          $ref: "#/components/schemas/AppraisalStatus"
...
    MyAppraisalSummary:
      type: object
      properties:
        employee_id:
          type: string
          example: EMP001
        my_appraisals_count:
          type: integer
          example: 5
        my_reviews_count:
          type: integer
          example: 7
        approved_count:
          type: integer
          example: 3
        submitted_reviews_count:
          type: integer
          example: 2
        average_rating:
          type: number
          format: float
          example: 4.2
    AppraisalHistoryResponse:
      type: object
      properties:
        count:
          type: integer
          example: 4
        employee_id:
          type: string
          example: EMP001
        appraisals:
          type: array
          items:
            $ref: "#/components/schemas/Appraisal"
    MyReviewQueueResponse:
      type: object
      properties:
        count:
          type: integer
          example: 2
        reviewer_id:
          type: string
          example: EMP002
        status:
          $ref: "#/components/schemas/AppraisalStatus"
        appraisals:
          type: array
          items:
            $ref: "#/components/schemas/Appraisal"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong

  # ── Reusable Responses ───────────────────────────────────────────────
  responses:
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"